chore: 调整代金券文件结构
