+++ src/Plugin/Wechat/V3/Marketing/Coupon/Stock/StartPlugin.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon;
+namespace Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\Stock;
 
 use Closure;
 use Yansongda\Artful\Contract\PluginInterface;
@@ -12,14 +12,13 @@ use Yansongda\Artful\Exception\ServiceNotFoundException;
 use Yansongda\Artful\Logger;
 use Yansongda\Artful\Rocket;
 use Yansongda\Pay\Exception\Exception;
-
 use function Yansongda\Pay\get_wechat_config;
 
 /**
- * @see https://pay.weixin.qq.com/docs/merchant/apis/cash-coupons/stock/restart-stock.html
- * @see https://pay.weixin.qq.com/docs/partner/apis/cash-coupons/stock/restart-stock.html
+ * @see https://pay.weixin.qq.com/docs/merchant/apis/cash-coupons/stock/start-stock.html
+ * @see https://pay.weixin.qq.com/docs/partner/apis/cash-coupons/stock/start-stock.html
  */
-class RestartPlugin implements PluginInterface
+class StartPlugin implements PluginInterface
 {
     /**
      * @throws ContainerException
@@ -28,13 +27,12 @@ class RestartPlugin implements PluginInterface
      */
     public function assembly(Rocket $rocket, Closure $next): Rocket
     {
-        Logger::debug('[Wechat][V3][Marketing][Coupon][RestartPlugin] 插件开始装载', ['rocket' => $rocket]);
+        Logger::debug('[Wechat][V3][Marketing][Coupon][Stock][StartPlugin] 插件开始装载', ['rocket' => $rocket]);
 
         $params = $rocket->getParams();
         $config = get_wechat_config($params);
         $payload = $rocket->getPayload();
         $stockId = $payload?->get('stock_id') ?? null;
-        $stockCreatorMchId = $payload?->get('stock_creator_mchid') ?? $config['mch_id'] ?? '';
 
         if (empty($stockId)) {
             throw new InvalidParamsException(Exception::PARAMS_NECESSARY_PARAMS_MISSING, '参数异常: 激活代金券，参数缺少 `stock_id`');
@@ -42,12 +40,12 @@ class RestartPlugin implements PluginInterface
 
         $rocket->setPayload([
             '_method' => 'POST',
-            '_url' => 'v3/marketing/favor/stocks/'.$stockId.'/restart',
-            '_service_url' => 'v3/marketing/favor/stocks/'.$stockId.'/restart',
-            'stock_creator_mchid' => $stockCreatorMchId,
+            '_url' => 'v3/marketing/favor/stocks/'.$stockId.'/start',
+            '_service_url' => 'v3/marketing/favor/stocks/'.$stockId.'/start',
+            'stock_creator_mchid' => $payload->get('stock_creator_mchid') ?? $config['mch_id'] ?? '',
         ]);
 
-        Logger::info('[Wechat][V3][Marketing][Coupon][RestartPlugin] 插件装载完毕', ['rocket' => $rocket]);
+        Logger::info('[Wechat][V3][Marketing][Coupon][Stock][StartPlugin] 插件装载完毕', ['rocket' => $rocket]);
 
         return $next($rocket);
     }
