+++ tests/Plugin/Wechat/V3/Marketing/Coupon/QueryDetailPluginTest.php
@@ -4,20 +4,20 @@ namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon;
 
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidParamsException;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\QueryStockDetailPlugin;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\QueryDetailPlugin;
 use Yansongda\Artful\Rocket;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-class QueryStockDetailPluginTest extends TestCase
+class QueryDetailPluginTest extends TestCase
 {
-    protected QueryStockDetailPlugin $plugin;
+    protected QueryDetailPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new QueryStockDetailPlugin();
+        $this->plugin = new QueryDetailPlugin();
     }
 
     public function testEmptyPayload()
@@ -26,7 +26,7 @@ class QueryStockDetailPluginTest extends TestCase
 
         self::expectException(InvalidParamsException::class);
         self::expectExceptionCode(Exception::PARAMS_NECESSARY_PARAMS_MISSING);
-        self::expectExceptionMessage('参数异常: 查询代金券批次详情，参数缺少 `stock_id`');
+        self::expectExceptionMessage('参数异常: 查询代金券详情，参数缺少 `openid` 或 `coupon_id`');
 
         $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
     }
@@ -35,16 +35,17 @@ class QueryStockDetailPluginTest extends TestCase
     {
         $rocket = new Rocket();
         $rocket->setPayload(new Collection( [
-            "stock_id" => "111",
-            'stock_creator_mchid' => '222',
+            "openid" => "111",
+            'coupon_id' => '222',
+            'appid' => '333',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
             '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=222',
-            '_service_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=222',
+            '_url' => 'v3/marketing/favor/users/111/coupons/222?appid=333',
+            '_service_url' => 'v3/marketing/favor/users/111/coupons/222?appid=333',
         ], $result->getPayload()->all());
     }
 
@@ -52,15 +53,16 @@ class QueryStockDetailPluginTest extends TestCase
     {
         $rocket = new Rocket();
         $rocket->setPayload(new Collection( [
-            "stock_id" => "111",
+            "openid" => "111",
+            'coupon_id' => '222',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
             '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=1600314069',
-            '_service_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=1600314069',
+            '_url' => 'v3/marketing/favor/users/111/coupons/222?appid=wx55955316af4ef13',
+            '_service_url' => 'v3/marketing/favor/users/111/coupons/222?appid=wx55955316af4ef13',
         ], $result->getPayload()->all());
     }
 }
