+++ tests/Plugin/Wechat/V3/Marketing/Coupon/Stock/QueryDetailPluginTest.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon;
+namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon\Stock;
 
-use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidParamsException;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\QueryStockItemsPlugin;
 use Yansongda\Artful\Rocket;
+use Yansongda\Pay\Exception\Exception;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\Stock\QueryDetailPlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-class QueryStockItemsPluginTest extends TestCase
+class QueryDetailPluginTest extends TestCase
 {
-    protected QueryStockItemsPlugin $plugin;
+    protected QueryDetailPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new QueryStockItemsPlugin();
+        $this->plugin = new QueryDetailPlugin();
     }
 
     public function testEmptyPayload()
@@ -26,7 +26,7 @@ class QueryStockItemsPluginTest extends TestCase
 
         self::expectException(InvalidParamsException::class);
         self::expectExceptionCode(Exception::PARAMS_NECESSARY_PARAMS_MISSING);
-        self::expectExceptionMessage('参数异常: 查询代金券可用单品，参数缺少 `stock_id`');
+        self::expectExceptionMessage('参数异常: 查询代金券批次详情，参数缺少 `stock_id`');
 
         $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
     }
@@ -34,18 +34,17 @@ class QueryStockItemsPluginTest extends TestCase
     public function testNormalParams()
     {
         $rocket = new Rocket();
-        $rocket->setPayload(new Collection([
+        $rocket->setPayload(new Collection( [
             "stock_id" => "111",
             'stock_creator_mchid' => '222',
-            '_t' => 'a',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
             '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=222',
-            '_service_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=222',
+            '_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=222',
+            '_service_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=222',
         ], $result->getPayload()->all());
     }
 
@@ -54,15 +53,14 @@ class QueryStockItemsPluginTest extends TestCase
         $rocket = new Rocket();
         $rocket->setPayload(new Collection( [
             "stock_id" => "111",
-            '_t' => 'a',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
             '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=1600314069',
-            '_service_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=1600314069',
+            '_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=1600314069',
+            '_service_url' => 'v3/marketing/favor/stocks/111?stock_creator_mchid=1600314069',
         ], $result->getPayload()->all());
     }
 }
