+++ tests/Plugin/Wechat/V3/Marketing/Coupon/Stock/QueryItemsPluginTest.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon;
+namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon\Stock;
 
-use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidParamsException;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\QueryStockMerchantsPlugin;
 use Yansongda\Artful\Rocket;
+use Yansongda\Pay\Exception\Exception;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\Stock\QueryItemsPlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-class QueryStockMerchantsPluginTest extends TestCase
+class QueryItemsPluginTest extends TestCase
 {
-    protected QueryStockMerchantsPlugin $plugin;
+    protected QueryItemsPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new QueryStockMerchantsPlugin();
+        $this->plugin = new QueryItemsPlugin();
     }
 
     public function testEmptyPayload()
@@ -26,7 +26,7 @@ class QueryStockMerchantsPluginTest extends TestCase
 
         self::expectException(InvalidParamsException::class);
         self::expectExceptionCode(Exception::PARAMS_NECESSARY_PARAMS_MISSING);
-        self::expectExceptionMessage('参数异常: 查询代金券可用商户，参数缺少 `stock_id`');
+        self::expectExceptionMessage('参数异常: 查询代金券可用单品，参数缺少 `stock_id`');
 
         $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
     }
@@ -34,7 +34,7 @@ class QueryStockMerchantsPluginTest extends TestCase
     public function testNormalParams()
     {
         $rocket = new Rocket();
-        $rocket->setPayload(new Collection( [
+        $rocket->setPayload(new Collection([
             "stock_id" => "111",
             'stock_creator_mchid' => '222',
             '_t' => 'a',
@@ -44,8 +44,8 @@ class QueryStockMerchantsPluginTest extends TestCase
 
         self::assertEquals([
             '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=222',
-            '_service_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=222',
+            '_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=222',
+            '_service_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=222',
         ], $result->getPayload()->all());
     }
 
@@ -61,8 +61,8 @@ class QueryStockMerchantsPluginTest extends TestCase
 
         self::assertEquals([
             '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=1600314069',
-            '_service_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=1600314069',
+            '_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=1600314069',
+            '_service_url' => 'v3/marketing/favor/stocks/111/items?stock_creator_mchid=1600314069',
         ], $result->getPayload()->all());
     }
 }
