+++ tests/Plugin/Wechat/V3/Marketing/Coupon/Stock/QueryMerchantsPluginTest.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon;
+namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon\Stock;
 
-use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidParamsException;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\PausePlugin;
 use Yansongda\Artful\Rocket;
+use Yansongda\Pay\Exception\Exception;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\Stock\QueryMerchantsPlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-class PausePluginTest extends TestCase
+class QueryMerchantsPluginTest extends TestCase
 {
-    protected PausePlugin $plugin;
+    protected QueryMerchantsPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new PausePlugin();
+        $this->plugin = new QueryMerchantsPlugin();
     }
 
     public function testEmptyPayload()
@@ -26,7 +26,7 @@ class PausePluginTest extends TestCase
 
         self::expectException(InvalidParamsException::class);
         self::expectExceptionCode(Exception::PARAMS_NECESSARY_PARAMS_MISSING);
-        self::expectExceptionMessage('参数异常: 暂停代金券批次，参数缺少 `stock_id`');
+        self::expectExceptionMessage('参数异常: 查询代金券可用商户，参数缺少 `stock_id`');
 
         $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
     }
@@ -35,17 +35,17 @@ class PausePluginTest extends TestCase
     {
         $rocket = new Rocket();
         $rocket->setPayload(new Collection( [
-            "stock_id" => "yansongda",
-            'stock_creator_mchid' => '1111',
+            "stock_id" => "111",
+            'stock_creator_mchid' => '222',
+            '_t' => 'a',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
-            '_method' => 'POST',
-            '_url' => 'v3/marketing/favor/stocks/yansongda/pause',
-            '_service_url' => 'v3/marketing/favor/stocks/yansongda/pause',
-            'stock_creator_mchid' => '1111',
+            '_method' => 'GET',
+            '_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=222',
+            '_service_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=222',
         ], $result->getPayload()->all());
     }
 
@@ -53,16 +53,16 @@ class PausePluginTest extends TestCase
     {
         $rocket = new Rocket();
         $rocket->setPayload(new Collection( [
-            "stock_id" => "yansongda",
+            "stock_id" => "111",
+            '_t' => 'a',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
-            '_method' => 'POST',
-            '_url' => 'v3/marketing/favor/stocks/yansongda/pause',
-            '_service_url' => 'v3/marketing/favor/stocks/yansongda/pause',
-            'stock_creator_mchid' => '1600314069',
+            '_method' => 'GET',
+            '_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=1600314069',
+            '_service_url' => 'v3/marketing/favor/stocks/111/merchants?stock_creator_mchid=1600314069',
         ], $result->getPayload()->all());
     }
 }
