+++ tests/Plugin/Wechat/V3/Marketing/Coupon/Stock/QueryRefundFlowPluginTest.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon;
+namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon\Stock;
 
-use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidParamsException;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\QueryStockRefundFlowPlugin;
 use Yansongda\Artful\Rocket;
+use Yansongda\Pay\Exception\Exception;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\Stock\QueryRefundFlowPlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-class QueryStockRefundFlowPluginTest extends TestCase
+class QueryRefundFlowPluginTest extends TestCase
 {
-    protected QueryStockRefundFlowPlugin $plugin;
+    protected QueryRefundFlowPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new QueryStockRefundFlowPlugin();
+        $this->plugin = new QueryRefundFlowPlugin();
     }
 
     public function testEmptyPayload()
