+++ tests/Plugin/Wechat/V3/Marketing/Coupon/Stock/RestartPluginTest.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon;
+namespace Yansongda\Pay\Tests\Plugin\Wechat\V3\Marketing\Coupon\Stock;
 
-use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidParamsException;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\QueryCouponDetailPlugin;
 use Yansongda\Artful\Rocket;
+use Yansongda\Pay\Exception\Exception;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Coupon\Stock\RestartPlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
-class QueryCouponDetailPluginTest extends TestCase
+class RestartPluginTest extends TestCase
 {
-    protected QueryCouponDetailPlugin $plugin;
+    protected RestartPlugin $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new QueryCouponDetailPlugin();
+        $this->plugin = new RestartPlugin();
     }
 
     public function testEmptyPayload()
@@ -26,7 +26,7 @@ class QueryCouponDetailPluginTest extends TestCase
 
         self::expectException(InvalidParamsException::class);
         self::expectExceptionCode(Exception::PARAMS_NECESSARY_PARAMS_MISSING);
-        self::expectExceptionMessage('参数异常: 查询代金券详情，参数缺少 `openid` 或 `coupon_id`');
+        self::expectExceptionMessage('参数异常: 激活代金券，参数缺少 `stock_id`');
 
         $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
     }
@@ -35,17 +35,17 @@ class QueryCouponDetailPluginTest extends TestCase
     {
         $rocket = new Rocket();
         $rocket->setPayload(new Collection( [
-            "openid" => "111",
-            'coupon_id' => '222',
-            'appid' => '333',
+            "stock_id" => "111",
+            'stock_creator_mchid' => '222',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
-            '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/users/111/coupons/222?appid=333',
-            '_service_url' => 'v3/marketing/favor/users/111/coupons/222?appid=333',
+            '_method' => 'POST',
+            '_url' => 'v3/marketing/favor/stocks/111/restart',
+            '_service_url' => 'v3/marketing/favor/stocks/111/restart',
+            'stock_creator_mchid' => '222',
         ], $result->getPayload()->all());
     }
 
@@ -53,16 +53,16 @@ class QueryCouponDetailPluginTest extends TestCase
     {
         $rocket = new Rocket();
         $rocket->setPayload(new Collection( [
-            "openid" => "111",
-            'coupon_id' => '222',
+            "stock_id" => "111",
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
 
         self::assertEquals([
-            '_method' => 'GET',
-            '_url' => 'v3/marketing/favor/users/111/coupons/222?appid=wx55955316af4ef13',
-            '_service_url' => 'v3/marketing/favor/users/111/coupons/222?appid=wx55955316af4ef13',
+            '_method' => 'POST',
+            '_url' => 'v3/marketing/favor/stocks/111/restart',
+            '_service_url' => 'v3/marketing/favor/stocks/111/restart',
+            'stock_creator_mchid' => '1600314069',
         ], $result->getPayload()->all());
     }
 }
